Moved program.cpp/hpp into opzioni.cpp/hpp
