+++ src/opzioni.cpp
@@ -1,6 +1,5 @@
-#include "program.hpp"
+#include "opzioni.hpp"
 
-#include <optional>
 #include <ranges>
 
 namespace opzioni {
